Switch to pactflow broker, split pacts for sandbox and live, verify all pacts using pact-js v3 verifier
