+++ tests/e2e/pact/Makefile
@@ -10,7 +10,11 @@ test:
 
 create-pacts:
 	rm -rf pact
-	npm run create-pacts
+ifeq (sandbox, $(findstring sandbox,$(APIGEE_ENVIRONMENT)))
+	npm run create-sandbox-pacts
+else
+	npm run create-live-pacts
+endif
 
 publish-pacts:
 	npm run publish-pacts
