+++ tests/e2e/pact/broker/verify.ts
@@ -1,28 +1,37 @@
 import { VerifierV3 } from "@pact-foundation/pact"
 
+let token: string
+
 /* eslint-disable  @typescript-eslint/no-explicit-any */
 async function verify(): Promise<any> { 
   const isLocal = process.env.PACT_PROVIDER_URL === "http://localhost:9000"
   const verifier =  new VerifierV3({
     publishVerificationResult: !isLocal,
-    pactBrokerUrl: isLocal ? undefined : process.env.PACT_BROKER_URL,
-    pactBrokerUsername: process.env.PACT_BROKER_BASIC_AUTH_USERNAME,
-    pactBrokerPassword: process.env.PACT_BROKER_BASIC_AUTH_PASSWORD,
+    pactBrokerUrl: isLocal ? undefined : process.env.PACT_BROKER_NEXT_URL,
+    pactBrokerToken: process.env.PACT_BROKER_NEXT_TOKEN,
     consumerVersionTag: process.env.PACT_VERSION,
     provider: `${process.env.PACT_PROVIDER}+${process.env.PACT_VERSION}`,
     providerVersion: process.env.PACT_VERSION,
     providerBaseUrl: process.env.PACT_PROVIDER_URL,
-    logLevel: isLocal? "debug" : "info",
+    logLevel: "debug",
+    stateHandlers: {
+      "is authenticated": () => {
+        token = `${process.env.APIGEE_ACCESS_TOKEN}`
+        Promise.resolve(`Valid bearer token generated`)
+      },
+      "is not authenticated": () => {
+        token = ""
+        Promise.resolve(`Invalid bearer token generated`)
+      }
+    },
     requestFilter: (req) => {
       req.headers["x-smoke-test"] = "1"
-      req.headers["Authorization"] = 
-        req.headers["Authorization"] 
-        ?? `Bearer ${process.env.APIGEE_ACCESS_TOKEN}`
+      req.headers["Authorization"] = `Bearer ${token}`
       return req
     },
     pactUrls: isLocal 
       ? [
-        `${process.cwd()}/pact/pacts/${process.env.PACT_CONSUMER}+${process.env.PACT_VERSION}-${process.env.PACT_PROVIDER}-convert+${process.env.PACT_VERSION}.json`
+        `${process.cwd()}/pact/pacts/${process.env.PACT_CONSUMER}+${process.env.PACT_VERSION}-${process.env.PACT_PROVIDER}+${process.env.PACT_VERSION}.json`
       ]
       : []
   })
@@ -31,6 +40,6 @@ async function verify(): Promise<any> {
 }
 
 (async () => {
-  verify()
+  verify().catch(verify).catch(verify)
 })()
