+++ tests/e2e/pact/specs/sandbox/eps-sandbox.process.spec.ts
@@ -1,0 +1,97 @@
+import {InteractionObject, Matchers} from "@pact-foundation/pact"
+import * as jestpact from "jest-pact"
+import supertest from "supertest"
+import * as TestResources from "../../resources/test-resources"
+import {Bundle} from "../../models/fhir/fhir-resources"
+import * as LosslessJson from "lossless-json"
+import {processExamples} from "../../services/process-example-fetcher"
+
+jestpact.pactWith(
+  {
+    spec: 3,
+    consumer: `nhsd-apim-eps-test-client+${process.env.PACT_VERSION}`,
+    provider: `nhsd-apim-eps-sandbox+${process.env.PACT_VERSION}`,
+    pactfileWriteMode: "merge"
+  },
+  /* eslint-disable  @typescript-eslint/no-explicit-any */
+  async (provider: any) => {
+    const client = () => {
+      const url = `${provider.mockService.baseUrl}`
+      return supertest(url)
+    }
+
+    describe("process-message sandbox e2e tests", () => {
+
+      const processCasesSubset = TestResources.processCases.splice(0, 5)
+
+      test.each(processCasesSubset)("should be able to process %s", async (desc: string, message: Bundle) => {
+        const apiPath = "/$process-message"
+        const messageStr = LosslessJson.stringify(message)
+        const interaction: InteractionObject = {
+          state: null,
+          uponReceiving: `a request to process ${desc} message to Spine`,
+          withRequest: {
+            headers: {
+              "Content-Type": "application/fhir+json; fhirVersion=4.0"
+            },
+            method: "POST",
+            path: "/$process-message",
+            body: JSON.parse(messageStr)
+          },
+          willRespondWith: {
+            headers: {
+              "Content-Type": "application/fhir+json; fhirVersion=4.0"
+            },
+            body: {
+              resourceType: "OperationOutcome",
+              issue: [
+                {
+                  code: Matchers.string("informational"),
+                  severity: Matchers.string("information")
+                }
+              ]
+            },
+            status: 200
+          }
+        }
+        await provider.addInteraction(interaction)
+        await client()
+          .post(apiPath)
+          .set('Content-Type', 'application/fhir+json; fhirVersion=4.0')
+          .send(messageStr)
+          .expect(200)
+      })
+
+      test("Should be able to process a FHIR JSON Accept header", async () => {
+        const testCase = processExamples[0]
+
+        const apiPath = "/$process-message"
+        const messageStr = LosslessJson.stringify(testCase.request)
+
+        const interaction: InteractionObject = {
+          state: null,
+          uponReceiving: `a request to process a message with a FHIR JSON Accept header`,
+          withRequest: {
+            headers: {
+              "Content-Type": "application/fhir+json; fhirVersion=4.0",
+              "Accept": "application/fhir+json"
+            },
+            method: "POST",
+            path: "/$process-message",
+            body: JSON.parse(messageStr)
+          },
+          willRespondWith: {
+            status: 200
+          }
+        }
+        await provider.addInteraction(interaction)
+        await client()
+          .post(apiPath)
+          .set('Content-Type', 'application/fhir+json; fhirVersion=4.0')
+          .set('Accept', 'application/fhir+json')
+          .send(messageStr)
+          .expect(200)
+      })
+    })
+  }
+)
